correct config type

+++ PlayerHeads-core/src/main/java/org/shininet/bukkit/playerheads/Config.java
@@ -74,10 +74,10 @@ public final class Config {
             put("fixdroppedheads",configType.BOOLEAN);
             
             
-            put("slimemodifier.0", configType.INT);
-            put("slimemodifier.1", configType.INT);
-            put("slimemodifier.2", configType.INT);
-            put("slimemodifier.3", configType.INT);
+            put("slimemodifier.0", configType.DOUBLE);
+            put("slimemodifier.1", configType.DOUBLE);
+            put("slimemodifier.2", configType.DOUBLE);
+            put("slimemodifier.3", configType.DOUBLE);
         }
     };
     /**
